requirements.txt
# Promptification MVP Requirements

# Core Framework
streamlit==1.29.0

# AI Integration
openai==1.6.1
anthropic==0.8.1

# Data Validation
pydantic==2.5.3

# Testing
pytest==7.4.3
pytest-mock==3.12.0

# Visualization
plotly==5.18.0
altair==5.2.0

# Utilities
python-dotenv==1.0.0
